requirements.txt
colorama==0.4.6
cryptography==46.0.3
h11==0.16.0
rich==14.2.0
textual==6.11.0
scapy==2.7.0
pyshark==0.6
pyOpenSSL==25.3.0